refactor: adminrow 분리
